// src/app/models/edit.ts
export class Edit {
	name: string;
    first_name: string;
    last_name: string;
    email: string;
    photo: File;
    id: string;
    city: any;
    state: string;
    country:string;
    gender: string;
	birthday: any;
    status: number = 1;
    is_updates: any;
    is_terms: any;
    is_age_validation: any;
    password: string;
    cpassword: string;
    auto_state:any;

}